El que hace pull es gay

+++ compra-venta/src/main/java/utn/frc/bda/dtos/NuevaOrdenCompraDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class NuevaOrdenCompraDTO {
     private String simbolo;
     private Long cantidad;
-    private Long montoMaximo;
+    private Double montoMaximo;
 }
